update and reorganize command line tool code
