+++ python/mcapi/cli/__init__.py
@@ -1,0 +1,3 @@
+from mcapi.cli.parser import CommonsCLIParser
+
+__all__ = dir()
