+++ python/mcapi/cli/init.py
@@ -1,0 +1,43 @@
+import os
+import sys
+import json
+import argparse
+import mcapi
+from mcapi.cli.functions import _mc_remotes, _print_projects, _proj_config
+
+def init_subcommand():
+        """
+        Initialize a new project
+        
+        mc init [--remote <remote>] [--desc <description>]
+        
+        """
+        parser = argparse.ArgumentParser(
+            description='Initialize current working directory as a new project',
+            prog='mc init')
+        parser.add_argument('--remote', type=str, default='origin', help='Remote name')
+        parser.add_argument('--desc', type=str, default='', help='Project description')
+        
+        # ignore 'mc init'
+        args = parser.parse_args(sys.argv[2:])
+        
+        remotes = _mc_remotes()
+        if not args.remote in remotes:
+            print "unrecognized remote:", args.remote
+            parser.print_help()
+            exit(1)
+        remote=remotes[args.remote]
+        
+        name = os.path.basename(os.getcwd())
+        
+        project = mcapi.create_project(name, args.desc)
+        
+        print "Created new project at:", remote.config.mcurl
+        _print_projects([project], project)
+        
+        if not os.path.exists('.mc'):
+            os.mkdir('.mc')
+        with open(_proj_config(), 'w') as f:
+            # set a default current experiment? 
+            data = {'remote_url':remote.config.mcurl, 'project_id': project.id, 'experiment_id': None}
+            json.dump(data, f)
