+++ python/mcapi/cli/remote.py
@@ -1,0 +1,30 @@
+import sys
+import argparse
+from mcapi.cli.functions import _mc_remotes
+import pandas
+
+def remote_subcommand():
+    """
+    Show / modify list of known Materials Commons servers.
+    
+    Goal:
+        mc remote [-v]
+        mc remote add 
+        mc remote remove
+    
+    Current:
+        mc remote
+    
+    """
+    parser = argparse.ArgumentParser(
+        description='Server settings',
+        prog='mc remote')
+    
+    # ignore 'mc remote'
+    args = parser.parse_args(sys.argv[2:])
+    
+    remotes = _mc_remotes()
+    data = [{'name': key, 'url': remotes[key].config.mcurl} for key in remotes]
+    df = pandas.DataFrame.from_records(data, columns=['name', 'url'])
+    print df.to_string(index=False)
+
