+++ python/scripts/mc
@@ -1,9 +1,9 @@
 #!/usr/bin/env python
 
 # sketch of a command line interface
-import mcapi.misc    
+import mcapi.cli   
 import types
 
 if __name__ == "__main__":
-    cli = mcapi.misc.CommonsCLIParser()
+    cli = mcapi.cli.CommonsCLIParser()
